Extract setSaveFileAs to the file
