+++ src/createMenu.ts
@@ -1,7 +1,6 @@
 import {
   BrowserWindow,
   dialog,
-  ipcMain,
   Menu,
   MenuItemConstructorOptions,
 } from "electron";
@@ -37,8 +36,8 @@ export const createMenu = (win: BrowserWindow): void => {
                 const fileData = {
                   status: true,
                   path: path,
-                  text: buff.toString()
-                }
+                  text: buff.toString(),
+                };
                 win.webContents.send("open-by-menu", fileData);
               })
               .catch((err) => console.log(`Error: ${err}`));
@@ -56,31 +55,4 @@ export const createMenu = (win: BrowserWindow): void => {
 
   const menu = Menu.buildFromTemplate(template);
   Menu.setApplicationMenu(menu);
-};
-
-export const setFileSaveAs = (win: BrowserWindow): void => {
-  ipcMain.handle("file-save-as", async (event, fileData) => {
-    return dialog
-      .showSaveDialog(win, {
-        properties: [
-          "showHiddenFiles",
-          "createDirectory",
-          "treatPackageAsDirectory",
-          "showOverwriteConfirmation",
-        ],
-        title: "ファイルを保存する",
-        filters: [
-          {
-            name: "Text",
-            extensions: ["txt", "text"],
-          },
-        ],
-      })
-      .then((result) => {
-        if (result.canceled) return;
-        const path = <fs.PathOrFileDescriptor>result.filePath;
-        fs.writeFileSync(path, fileData);
-      })
-      .catch((err) => console.log(`Error: ${err}`));
-  });
 };
